LayoutTests/imported/w3c/web-platform-tests/svg/text/scripted/getextentofchar-expected.txt

abcd
abc

PASS Multiple chunks in a tspan
FAIL With dominant-baseline assert_approx_equals: expected 147.5 +/- 1 but got 136
